+++ server/TWS Foundation/TWS Business/Depots/ApproachesHDepot.cs
@@ -1,4 +1,5 @@
 ﻿using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -11,6 +12,9 @@ public class ApproachesHDepot : BDepot<TWSBusinessDatabase, ApproachesH> {
     /// <summary>
     ///     Generates a new depot handler for <see cref="ApproachesH"/>.
     /// </summary>
+    public ApproachesHDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+      : base(Databases, Disposer) {
+    }
     public ApproachesHDepot() : base(new(), null) {
     }
 }