+++ server/TWS Foundation/TWS Business/Depots/DriversDepot.cs
@@ -1,4 +1,5 @@
 ﻿using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -11,6 +12,9 @@ public class DriversDepot : BDepot<TWSBusinessDatabase, Driver> {
     /// <summary>
     ///     Generates a new depot handler for <see cref="Driver"/>.
     /// </summary>
+    public DriversDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+        : base(Databases, Disposer) {
+    }
     public DriversDepot() : base(new(), null) {
     }
 }