+++ server/TWS Foundation/TWS Business/Depots/ManufacturersDepot.cs
@@ -1,5 +1,6 @@
 ﻿
 using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -12,6 +13,9 @@ public class ManufacturersDepot : BDepot<TWSBusinessDatabase, Manufacturer> {
     /// <summary>
     ///     Generates a new depot handler for <see cref="Manufacturer"/>.
     /// </summary>
+    public ManufacturersDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+        : base(Databases, Disposer) {
+    }
     public ManufacturersDepot() : base(new(), null) {
 
     }