+++ server/TWS Foundation/TWS Business/Depots/SituationsDepot.cs
@@ -1,4 +1,5 @@
 ﻿using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -13,6 +14,9 @@ public class SituationsDepot
     /// <summary>
     ///     Generates a new depot handler for <see cref="Situation"/>.
     /// </summary>
+    public SituationsDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+        : base(Databases, Disposer) {
+    }
     public SituationsDepot() : base(new(), null) {
 
     }