+++ server/TWS Foundation/TWS Business/Depots/TruckDepot.cs
@@ -1,5 +1,6 @@
 ﻿
 using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -12,6 +13,9 @@ public class TruckDepot : BDepot<TWSBusinessDatabase, Truck> {
     /// <summary>
     ///     Generates a new depot handler for <see cref="Truck"/>.
     /// </summary>
+    public TruckDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+       : base(Databases, Disposer) {
+    }
     public TruckDepot()
         : base(new(), null) {
     }