+++ server/TWS Foundation/TWS Business/Depots/VehiculesModelsDepot.cs
@@ -1,4 +1,5 @@
 ﻿using CSM_Foundation.Database.Bases;
+using CSM_Foundation.Database.Interfaces;
 
 using TWS_Business.Sets;
 
@@ -11,6 +12,9 @@ public class VehiculesModelsDepot : BDepot<TWSBusinessDatabase, VehiculeModel> {
     /// <summary>
     ///     Generates a new depot handler for <see cref="VehiculeModel"/>.
     /// </summary>
+    public VehiculesModelsDepot(TWSBusinessDatabase Databases, IDisposer? Disposer = null)
+       : base(Databases, Disposer) {
+    }
     public VehiculesModelsDepot() : base(new(), null) {
     }
 }