+++ server/TWS Foundation/TWS Foundation/Controllers/Business/VehiculesModelsController.cs
@@ -1,0 +1,30 @@
+﻿using CSM_Foundation.Database.Models.Options;
+
+using Microsoft.AspNetCore.Mvc;
+
+using TWS_Business.Sets;
+
+using TWS_Customer.Services.Interfaces;
+
+using TWS_Foundation.Authentication;
+
+namespace TWS_Foundation.Controllers.Business;
+
+[ApiController, Route("[Controller]")]
+public class VehiculesModelsController : ControllerBase {
+    private readonly IVehiculesModelsService Service;
+
+    public VehiculesModelsController(IVehiculesModelsService Service) {
+        this.Service = Service;
+    }
+
+    [HttpPost("[Action]"), Auth(["ABC1", "ABC2"])]
+    public async Task<IActionResult> View(SetViewOptions<VehiculeModel> Options) {
+        return Ok(await Service.View(Options));
+    }
+
+    [HttpPost("[Action]"), Auth(["ABC1", "ABC2"])]
+    public async Task<IActionResult> Create(VehiculeModel vehiculeModel) {
+        return Ok(await Service.Create(vehiculeModel));
+    }
+}